ref: Move packet building into conn
As suggested by @ramfox in https://github.com/n0-computer/iroh/pull/830#discussion_r1191830689
