+++ src/hp/derp/client_conn.rs
@@ -734,8 +734,6 @@ fn parse_send_packet(data: &[u8]) -> Result<(PublicKey, &[u8])> {
 
 #[cfg(test)]
 mod tests {
-    use crate::hp::derp::client::PacketSplitIter;
-
     use super::*;
     use anyhow::bail;
     use std::sync::Arc;
@@ -808,14 +806,13 @@ mod tests {
 
         // send packet
         let data = b"hello world!";
-        crate::hp::derp::client::send_packets(&mut writer, &None, target.clone(), &[data]).await?;
+        crate::hp::derp::client::send_packet(&mut writer, &None, target.clone(), &data[..]).await?;
         let msg = server_channel_r.recv().await.unwrap();
         match msg {
             ServerMessage::SendPacket((got_target, packet)) => {
-                let payload = PacketSplitIter::split(packet.bytes)?;
                 assert_eq!(target, got_target);
                 assert_eq!(key, packet.src);
-                assert_eq!(&data[..], &payload[0]);
+                assert_eq!(&data[..], &packet.bytes);
             }
             m => {
                 bail!("expected ServerMessage::SendPacket, got {m:?}");
@@ -827,7 +824,7 @@ mod tests {
         let mut disco_data = crate::hp::disco::MAGIC.as_bytes().to_vec();
         disco_data.extend_from_slice(target.as_bytes());
         disco_data.extend_from_slice(data);
-        crate::hp::derp::client::send_packets(&mut writer, &None, target.clone(), &[&disco_data])
+        crate::hp::derp::client::send_packet(&mut writer, &None, target.clone(), &disco_data)
             .await?;
         let msg = server_channel_r.recv().await.unwrap();
         match msg {
