+++ src/hp/derp/http.rs
@@ -23,7 +23,6 @@ mod tests {
     use tokio_util::sync::CancellationToken;
     use tracing_subscriber::{prelude::*, EnvFilter};
 
-    use crate::hp::derp::client::PacketSplitIter;
     use crate::hp::derp::{DerpNode, DerpRegion, ReceivedMessage, UseIpv4, UseIpv6};
     use crate::hp::{
         derp::Server as DerpServer,
@@ -131,7 +130,7 @@ mod tests {
 
         println!("sending message from a to b");
         let msg = Bytes::from_static(b"hi there, client b!");
-        client_a.send(b_key.clone(), vec![msg.clone()]).await?;
+        client_a.send(b_key.clone(), msg.clone()).await?;
         println!("waiting for message from a on b");
         let (got_key, got_msg) = b_recv.recv().await.expect("expected message from client_a");
         assert_eq!(a_key, got_key);
@@ -139,7 +138,7 @@ mod tests {
 
         println!("sending message from b to a");
         let msg = Bytes::from_static(b"right back at ya, client b!");
-        client_b.send(a_key.clone(), vec![msg.clone()]).await?;
+        client_b.send(a_key.clone(), msg.clone()).await?;
         println!("waiting for message b on a");
         let (got_key, got_msg) = a_recv.recv().await.expect("expected message from client_b");
         assert_eq!(b_key, got_key);
@@ -181,23 +180,16 @@ mod tests {
                     Ok((msg, _)) => {
                         println!("got message on {:?}: {msg:?}", key.public_key());
                         if let ReceivedMessage::ReceivedPacket { source, data } = msg {
-                            let iter = PacketSplitIter::new(data);
-                            for packet in iter {
-                                let Ok(data) = packet else {
-                                    tracing::warn!("error parsing packet");
-                                    return;
-                                };
-                                received_msg_s
-                                    .send((source.clone(), data))
-                                    .await
-                                    .unwrap_or_else(|err| {
-                                        panic!(
-                                            "client {:?}, error sending message over channel: {:?}",
-                                            key.public_key(),
-                                            err
-                                        )
-                                    });
-                            }
+                            received_msg_s
+                                .send((source.clone(), data))
+                                .await
+                                .unwrap_or_else(|err| {
+                                    panic!(
+                                        "client {:?}, error sending message over channel: {:?}",
+                                        key.public_key(),
+                                        err
+                                    )
+                                });
                         }
                     }
                 }
