+++ src/hp/derp/http/client.rs
@@ -595,11 +595,7 @@ impl Client {
     ///
     /// If there is an error sending the packet, it closes the underlying derp connection before
     /// returning.
-    pub async fn send(
-        &self,
-        dst_key: key::node::PublicKey,
-        b: Vec<Bytes>,
-    ) -> Result<(), ClientError> {
+    pub async fn send(&self, dst_key: key::node::PublicKey, b: Bytes) -> Result<(), ClientError> {
         debug!("send");
         let (client, _) = self.connect().await?;
         if client.send(dst_key, b).await.is_err() {
