+++ src/hp/derp/server.rs
@@ -575,9 +575,7 @@ mod tests {
 
     use crate::hp::{
         derp::{
-            client::{ClientBuilder, PacketSplitIter},
-            client_conn::ClientConnBuilder,
-            types::ClientInfo,
+            client::ClientBuilder, client_conn::ClientConnBuilder, types::ClientInfo,
             ReceivedMessage, MAX_FRAME_SIZE,
         },
         key::node::PUBLIC_KEY_LENGTH,
@@ -730,22 +728,21 @@ mod tests {
 
         // write message from b to a
         let msg = b"hello world!";
-        crate::hp::derp::client::send_packets(&mut b_writer, &None, key_a.clone(), &[msg]).await?;
+        crate::hp::derp::client::send_packet(&mut b_writer, &None, key_a.clone(), &msg[..]).await?;
 
         // get message on a's reader
         let (frame_type, _) =
             crate::hp::derp::read_frame(&mut a_reader, MAX_FRAME_SIZE, &mut buf).await?;
         let (key, frame) = crate::hp::derp::client::parse_recv_frame(buf.clone())?;
-        let payload = PacketSplitIter::split(frame)?;
         assert_eq!(FrameType::RecvPacket, frame_type);
         assert_eq!(key_b, key);
-        assert_eq!(msg, &payload[0][..]);
+        assert_eq!(msg, &frame[..]);
 
         // write disco message from b to d
         let mut disco_msg = crate::hp::disco::MAGIC.as_bytes().to_vec();
         disco_msg.extend_from_slice(key_b.as_bytes());
         disco_msg.extend_from_slice(msg);
-        crate::hp::derp::client::send_packets(&mut b_writer, &None, key_d.clone(), &[&disco_msg])
+        crate::hp::derp::client::send_packet(&mut b_writer, &None, key_d.clone(), &disco_msg)
             .await?;
 
         // get message on d's reader
@@ -905,14 +902,11 @@ mod tests {
 
         // send message from a to b!
         let msg = Bytes::from_static(b"hello client b!!");
-        client_a
-            .send(public_key_b.clone(), vec![msg.clone()])
-            .await?;
+        client_a.send(public_key_b.clone(), msg.clone()).await?;
         match client_b.recv().await? {
             ReceivedMessage::ReceivedPacket { source, data } => {
-                let payload = PacketSplitIter::split(data)?;
                 assert_eq!(public_key_a, source);
-                assert_eq!(&msg[..], payload[0]);
+                assert_eq!(&msg[..], data);
             }
             msg => {
                 anyhow::bail!("expected ReceivedPacket msg, got {msg:?}");
@@ -921,14 +915,11 @@ mod tests {
 
         // send message from b to a!
         let msg = Bytes::from_static(b"nice to meet you client a!!");
-        client_b
-            .send(public_key_a.clone(), vec![msg.clone()])
-            .await?;
+        client_b.send(public_key_a.clone(), msg.clone()).await?;
         match client_a.recv().await? {
             ReceivedMessage::ReceivedPacket { source, data } => {
-                let payload = PacketSplitIter::split(data)?;
                 assert_eq!(public_key_b, source);
-                assert_eq!(&msg[..], payload[0]);
+                assert_eq!(&msg[..], data);
             }
             msg => {
                 anyhow::bail!("expected ReceivedPacket msg, got {msg:?}");
@@ -937,18 +928,17 @@ mod tests {
 
         // send message from a to c
         let msg = Bytes::from_static(b"can you pass this to client d?");
-        client_a.send(key_c.clone(), vec![msg.clone()]).await?;
+        client_a.send(key_c.clone(), msg.clone()).await?;
         let (got_src, got_dst, got_packet) = fwd_recv.recv().await.unwrap();
-        let payload = PacketSplitIter::split(got_packet)?;
         assert_eq!(public_key_a, got_src);
         assert_eq!(key_c, got_dst);
-        assert_eq!(&msg[..], payload[0]);
+        assert_eq!(&msg[..], got_packet);
 
         // remove the packet forwarder for c
         handler.remove_packet_forwarder(key_c.clone())?;
         // try to send c a message
         let msg = Bytes::from_static(b"can you pass this to client d?");
-        client_a.send(key_c, vec![msg.clone()]).await?;
+        client_a.send(key_c, msg.clone()).await?;
         // packet forwarder has been removed
         assert!(fwd_recv.recv().await.is_none());
 
@@ -957,7 +947,7 @@ mod tests {
 
         // client connections have been shutdown
         client_a
-            .send(public_key_b, vec![Bytes::from_static(b"try to send")])
+            .send(public_key_b, Bytes::from_static(b"try to send"))
             .await?;
         assert!(client_b.recv().await.is_err());
         Ok(())
@@ -997,14 +987,11 @@ mod tests {
 
         // send message from a to b!
         let msg = Bytes::from_static(b"hello client b!!");
-        client_a
-            .send(public_key_b.clone(), vec![msg.clone()])
-            .await?;
+        client_a.send(public_key_b.clone(), msg.clone()).await?;
         match client_b.recv().await? {
             ReceivedMessage::ReceivedPacket { source, data } => {
-                let payload = PacketSplitIter::split(data)?;
                 assert_eq!(public_key_a, source);
-                assert_eq!(&msg[..], payload[0]);
+                assert_eq!(&msg[..], data);
             }
             msg => {
                 anyhow::bail!("expected ReceivedPacket msg, got {msg:?}");
@@ -1013,14 +1000,11 @@ mod tests {
 
         // send message from b to a!
         let msg = Bytes::from_static(b"nice to meet you client a!!");
-        client_b
-            .send(public_key_a.clone(), vec![msg.clone()])
-            .await?;
+        client_b.send(public_key_a.clone(), msg.clone()).await?;
         match client_a.recv().await? {
             ReceivedMessage::ReceivedPacket { source, data } => {
-                let payload = PacketSplitIter::split(data)?;
                 assert_eq!(public_key_b, source);
-                assert_eq!(&msg[..], payload[0]);
+                assert_eq!(&msg[..], data);
             }
             msg => {
                 anyhow::bail!("expected ReceivedPacket msg, got {msg:?}");
@@ -1039,14 +1023,11 @@ mod tests {
 
         // send message from a to b!
         let msg = Bytes::from_static(b"are you still there, b?!");
-        client_a
-            .send(public_key_b.clone(), vec![msg.clone()])
-            .await?;
+        client_a.send(public_key_b.clone(), msg.clone()).await?;
         match new_client_b.recv().await? {
             ReceivedMessage::ReceivedPacket { source, data } => {
-                let payload = PacketSplitIter::split(data)?;
                 assert_eq!(public_key_a, source);
-                assert_eq!(&msg[..], payload[0]);
+                assert_eq!(&msg[..], data);
             }
             msg => {
                 anyhow::bail!("expected ReceivedPacket msg, got {msg:?}");
@@ -1055,14 +1036,11 @@ mod tests {
 
         // send message from b to a!
         let msg = Bytes::from_static(b"just had a spot of trouble but I'm back now,a!!");
-        new_client_b
-            .send(public_key_a.clone(), vec![msg.clone()])
-            .await?;
+        new_client_b.send(public_key_a.clone(), msg.clone()).await?;
         match client_a.recv().await? {
             ReceivedMessage::ReceivedPacket { source, data } => {
-                let payload = PacketSplitIter::split(data)?;
                 assert_eq!(public_key_b, source);
-                assert_eq!(&msg[..], payload[0]);
+                assert_eq!(&msg[..], data);
             }
             msg => {
                 anyhow::bail!("expected ReceivedPacket msg, got {msg:?}");
@@ -1074,7 +1052,7 @@ mod tests {
 
         // client connections have been shutdown
         client_a
-            .send(public_key_b, vec![Bytes::from_static(b"try to send")])
+            .send(public_key_b, Bytes::from_static(b"try to send"))
             .await?;
         assert!(new_client_b.recv().await.is_err());
         Ok(())
